feat(item): refactor price type from double to int for database compatibility

+++ lib/features/menu/domain/entities/item.dart
@@ -1,10 +1,12 @@
+// مسار الملف: lib/features/menu/domain/entities/item.dart
+
 import 'package:equatable/equatable.dart';
 
 class Item extends Equatable {
   final int? id;
   final int categoryId;
   final String name;
-  final double price;
+  final int price; // Refactored: تغيير من double إلى int (Cents)
   final String? imagePath;
   final String createdAt;
   final String updatedAt;
